[DT-353] Refactor categories repository
