+++ feature_categories/src/main/AndroidManifest.xml
