+++ feature_categories/src/main/java/cm/aptoide/pt/feature_categories/data/model/CategoryJson.kt
