+++ feature_categories/src/main/java/cm/aptoide/pt/feature_categories/presentation/ViewModelProvider.kt
@@ -5,17 +5,13 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.viewModel
-import cm.aptoide.pt.feature_categories.domain.usecase.GetCategoriesListUseCase
+import cm.aptoide.pt.feature_categories.data.CategoriesRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
-interface CategoriesDependenciesProvider {
-  val getCategoriesListUseCase: GetCategoriesListUseCase
-}
-
 @HiltViewModel
 class InjectionsProvider @Inject constructor(
-  val provider: CategoriesDependenciesProvider,
+  val categoriesRepository: CategoriesRepository,
 ) : ViewModel()
 
 @Composable
@@ -28,7 +24,7 @@ fun CategoriesViewModel(requestUrl: String): CategoriesViewModel {
         @Suppress("UNCHECKED_CAST")
         return CategoriesViewModel(
           categoriesWidgetUrl = requestUrl,
-          getCategoriesListUseCase = injectionsProvider.provider.getCategoriesListUseCase,
+          categoriesRepository = injectionsProvider.categoriesRepository,
         ) as T
       }
     }
