--- feature_categories/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-  <string name="app_name">feature_categories</string>
-</resources>
